feature: 完善数据范围策略模式基于全部和自定义范围的扩展

+++ mate-platform/mate-system-api/src/main/java/vip/mate/system/dto/RoleDTO.java
@@ -13,6 +13,7 @@ public class RoleDTO {
 	private String roleDesc;
 	private String delFlag;
 	private int dsType;
-	List<Integer> permissionIds;
-	List<Integer> roleDepts;
+	List<Long> permissionIds;
+	List<Long> roleDepts;
+
 }
